libs/observability/pyproject.toml
[project]
name = "greyeye-observability"
version = "0.1.0"
description = "GreyEye shared observability — structured logging, Prometheus metrics, OpenTelemetry tracing"
requires-python = ">=3.11"

dependencies = [
    "structlog>=24.4",
    "prometheus-client>=0.21",
    "opentelemetry-api>=1.28",
    "opentelemetry-sdk>=1.28",
    "opentelemetry-exporter-otlp-proto-grpc>=1.28",
    "opentelemetry-instrumentation-fastapi>=0.49b0",
    "opentelemetry-instrumentation-httpx>=0.49b0",
    "opentelemetry-instrumentation-redis>=0.49b0",
    "opentelemetry-instrumentation-sqlalchemy>=0.49b0",
    "opentelemetry-instrumentation-logging>=0.49b0",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "fastapi>=0.115",
    "httpx>=0.27",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["observability"]

[tool.uv.sources]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
